fixed nonfavourites

+++ src/pl/edu/agh/fis/vtaskmaster/VirtualTaskmaster.java
@@ -347,7 +347,16 @@ public class VirtualTaskmaster {
         int selRow = vTMW.tblToDo.getSelectedRow();
         if (selRow != -1 && vTMW.tblToDo.getValueAt(selRow, 0) != null) {
             if(!database.getTaskByName((String) vTMW.tblToDo.getValueAt(selRow, 0)).isFavourite()){
-                database.removeTaskByName((String) vTMW.tblToDo.getValueAt(selRow, 0));
+                boolean inHistory = false;
+                ArrayList<Task> history = database.getHistory();
+                for(int i=0; i < history.size(); i++){
+                    if(history.get(i).getName().equals((String) vTMW.tblToDo.getValueAt(selRow, 0))){
+                        inHistory = true;
+                    }
+                }
+                if(!inHistory) {
+                    database.removeTaskByName((String) vTMW.tblToDo.getValueAt(selRow, 0));
+                }
             } else {
                 Task why = database.getTaskByName((String) vTMW.tblToDo.getValueAt(selRow, 0));
                 database.removeTaskByName((String) vTMW.tblToDo.getValueAt(selRow, 0));
@@ -442,18 +451,27 @@ public class VirtualTaskmaster {
         }else if(vTM.rS == returnState.VTM_TODO){
             if(validateDataVTM((int)vTM.spnr_hour.getValue(),(int)vTM.spnr_mint.getValue(),vTM.textField.getText(), vTM.textPane.getText())){
                 int eRow = tblFindEmptyRow(vTMW.tblToDo);
-                ((DefaultTableModel)vTMW.tblToDo.getModel()).addRow(new Object[]{null,null,null,null});
-                vTMW.tblToDo.setValueAt(vTM.textField.getText(), eRow, 0);
-                vTMW.tblToDo.setValueAt((int)vTM.spnr_prior.getValue(), eRow, 1);
-                vTMW.tblToDo.setValueAt(VTMainWindow.timeFiller((int)vTM.spnr_hour.getValue())+":"+ VTMainWindow.timeFiller((int)vTM.spnr_mint.getValue()), eRow, 2);
-                vTMW.tblToDo.setValueAt("00:00", eRow, 3);
+
                 if(database.getTaskByName(vTM.textField.getText()) == null){
+                    ((DefaultTableModel)vTMW.tblToDo.getModel()).addRow(new Object[]{null,null,null,null});
+                    vTMW.tblToDo.setValueAt(vTM.textField.getText(), eRow, 0);
+                    vTMW.tblToDo.setValueAt((int)vTM.spnr_prior.getValue(), eRow, 1);
+                    vTMW.tblToDo.setValueAt(VTMainWindow.timeFiller((int)vTM.spnr_hour.getValue())+":"+ VTMainWindow.timeFiller((int)vTM.spnr_mint.getValue()), eRow, 2);
+                    vTMW.tblToDo.setValueAt("00:00", eRow, 3);
                     database.saveTask(new Task(vTM.textField.getText(), vTM.textPane.getText(), (int)vTM.spnr_prior.getValue(), (long) ((Integer)vTM.spnr_hour.getValue()*3600000 + (Integer)vTM.spnr_mint.getValue()*60000), true, false));
+                } else if(database.getTaskByName(vTM.textField.getText()).isTodo()) {
+                    JOptionPane.showMessageDialog(new JFrame(), "This task is already set as TODO");
+                } else {
+                    ((DefaultTableModel)vTMW.tblToDo.getModel()).addRow(new Object[]{null,null,null,null});
+                    vTMW.tblToDo.setValueAt(vTM.textField.getText(), eRow, 0);
+                    vTMW.tblToDo.setValueAt((int)vTM.spnr_prior.getValue(), eRow, 1);
+                    vTMW.tblToDo.setValueAt(VTMainWindow.timeFiller((int)vTM.spnr_hour.getValue())+":"+ VTMainWindow.timeFiller((int)vTM.spnr_mint.getValue()), eRow, 2);
+                    vTMW.tblToDo.setValueAt("00:00", eRow, 3);
+                    Task why = database.getTaskByName(vTM.textField.getText());
+                    database.removeTaskByName(vTM.textField.getText());
+                    why.setTodo(true);
+                    database.saveTask(why);
                 }
-                Task why = database.getTaskByName(vTM.textField.getText());
-                database.removeTaskByName(vTM.textField.getText());
-                why.setTodo(true);
-                database.saveTask(why);
                 VTMainWindowManageTasksButton();
             }else{
                 JOptionPane.showMessageDialog(new JFrame(), "You have to provide full description of your task.");
@@ -474,7 +492,16 @@ public class VirtualTaskmaster {
         int selRow = tbl.getSelectedRow();
         if (selRow != -1 && tbl.getValueAt(selRow, 0) != null) {
             if(!database.getTaskByName((String) tbl.getValueAt(selRow, 0)).isTodo()){
-                database.removeTaskByName((String) tbl.getValueAt(selRow, 0));
+                boolean inHistory = false;
+                ArrayList<Task> history = database.getHistory();
+                for(int i=0; i < history.size(); i++){
+                    if(history.get(i).getName().equals((String) tbl.getValueAt(selRow, 0))){
+                        inHistory = true;
+                    }
+                }
+                if(!inHistory) {
+                    database.removeTaskByName((String) tbl.getValueAt(selRow, 0));
+                }
             } else {
                 Task why = database.getTaskByName((String) tbl.getValueAt(selRow, 0));
                 database.removeTaskByName((String) tbl.getValueAt(selRow, 0));
@@ -525,39 +552,71 @@ public class VirtualTaskmaster {
         vTM.tabEdit = true;
         if(selRow == -1)
             selRow = tblFindEmptyRow(tbl);
-        if(tbl.getValueAt(selRow, 0) == null){
-            ((DefaultTableModel) tbl.getModel()).addRow(new Object[]{null,null,null,null});
-        }
-        tbl.setValueAt(vTM.textField.getText(),selRow,0);
-        tbl.setValueAt(vTM.spnr_prior.getValue(),selRow,1);
-        if((int)(vTM.spnr_mint.getValue()) > 9){
-            tbl.setValueAt(vTM.spnr_hour.getValue()+":"+vTM.spnr_mint.getValue(),selRow,2);
-        }
-        else{
-            tbl.setValueAt(vTM.spnr_hour.getValue()+":0"+vTM.spnr_mint.getValue(),selRow,2);
-        }
-        try {
-            long averageTime = database.stats.averageTimeForTaskWithName(vTM.textField.getText());
-            int avrTimeH = (int) averageTime / 3600000; String ath;
-            int avrTimeM = (int) ((averageTime - avrTimeH * 3600000) / 60000); String atm;
-            if (avrTimeH < 10) ath = 0 + "" + avrTimeH;
-            else ath = "" + avrTimeH;
-            if (avrTimeM < 10) atm = 0 + "" + avrTimeM;
-            else atm = "" + avrTimeM;
 
-            tbl.setValueAt(ath + ":" + atm, selRow, 3);
-        } catch(SQLException exception) {
-            tbl.setValueAt("0:00", selRow, 3);
-        }
         vTM.tabEdit = false;
         System.out.println(vTM.textField.getText());
         if (tbl == vTM.tblFavourites) {
+            if(tbl.getValueAt(selRow, 0) == null){
+                ((DefaultTableModel) tbl.getModel()).addRow(new Object[]{null,null,null,null});
+            }
+            tbl.setValueAt(vTM.textField.getText(),selRow,0);
+            tbl.setValueAt(vTM.spnr_prior.getValue(),selRow,1);
+            if((int)(vTM.spnr_mint.getValue()) > 9){
+                tbl.setValueAt(vTM.spnr_hour.getValue()+":"+vTM.spnr_mint.getValue(),selRow,2);
+            }
+            else{
+                tbl.setValueAt(vTM.spnr_hour.getValue()+":0"+vTM.spnr_mint.getValue(),selRow,2);
+            }
+            try {
+                long averageTime = database.stats.averageTimeForTaskWithName(vTM.textField.getText());
+                int avrTimeH = (int) averageTime / 3600000; String ath;
+                int avrTimeM = (int) ((averageTime - avrTimeH * 3600000) / 60000); String atm;
+                if (avrTimeH < 10) ath = 0 + "" + avrTimeH;
+                else ath = "" + avrTimeH;
+                if (avrTimeM < 10) atm = 0 + "" + avrTimeM;
+                else atm = "" + avrTimeM;
+
+                tbl.setValueAt(ath + ":" + atm, selRow, 3);
+            } catch(SQLException exception) {
+                tbl.setValueAt("0:00", selRow, 3);
+            }
             if(database.getTaskByName((String) vTM.textField.getText()) == null){
                 database.saveTask(new Task(vTM.textField.getText(), vTM.textPane.getText(), (int)vTM.spnr_prior.getValue(), (long) ((Integer)vTM.spnr_hour.getValue()*3600000 + (Integer)vTM.spnr_mint.getValue()*60000), true, false));
+            } else {
+                Task why = database.getTaskByName(vTM.textField.getText());
+                database.removeTaskByName(vTM.textField.getText());
+                why.setFavourite(true);
+                database.saveTask(why);
             }
         } else {
             if(database.getTaskByName((String) vTM.textField.getText()) == null){
+                if(tbl.getValueAt(selRow, 0) == null){
+                    ((DefaultTableModel) tbl.getModel()).addRow(new Object[]{null,null,null,null});
+                }
+                tbl.setValueAt(vTM.textField.getText(),selRow,0);
+                tbl.setValueAt(vTM.spnr_prior.getValue(),selRow,1);
+                if((int)(vTM.spnr_mint.getValue()) > 9){
+                    tbl.setValueAt(vTM.spnr_hour.getValue()+":"+vTM.spnr_mint.getValue(),selRow,2);
+                }
+                else{
+                    tbl.setValueAt(vTM.spnr_hour.getValue()+":0"+vTM.spnr_mint.getValue(),selRow,2);
+                }
+                try {
+                    long averageTime = database.stats.averageTimeForTaskWithName(vTM.textField.getText());
+                    int avrTimeH = (int) averageTime / 3600000; String ath;
+                    int avrTimeM = (int) ((averageTime - avrTimeH * 3600000) / 60000); String atm;
+                    if (avrTimeH < 10) ath = 0 + "" + avrTimeH;
+                    else ath = "" + avrTimeH;
+                    if (avrTimeM < 10) atm = 0 + "" + avrTimeM;
+                    else atm = "" + avrTimeM;
+
+                    tbl.setValueAt(ath + ":" + atm, selRow, 3);
+                } catch(SQLException exception) {
+                    tbl.setValueAt("0:00", selRow, 3);
+                }
                 database.saveTask(new Task(vTM.textField.getText(), vTM.textPane.getText(), (int)vTM.spnr_prior.getValue(), (long) ((Integer)vTM.spnr_hour.getValue()*3600000 + (Integer)vTM.spnr_mint.getValue()*60000), false, false));
+            } else {
+                JOptionPane.showMessageDialog(new JFrame(), "Task with this name already exists in database");
             }
         }
     }
